Removed the 2^16 limit from TIFFCheckDirOffset

+++ libtiff/tiffiop.h
@@ -118,8 +118,8 @@ struct tiff {
 	uint64_t               tif_diroff;       /* file offset of current directory */
 	uint64_t               tif_nextdiroff;   /* file offset of following directory */
 	uint64_t*              tif_dirlist;      /* list of offsets to already seen directories to prevent IFD looping */
-	uint16_t               tif_dirlistsize;  /* number of entries in offset list */
-	uint16_t               tif_dirnumber;    /* number of already seen directories */
+    uint32_t               tif_dirlistsize;  /* number of entries in offset list */
+    uint32_t               tif_dirnumber;    /* number of already seen directories */
 	TIFFDirectory        tif_dir;          /* internal rep of current directory */
 	TIFFDirectory        tif_customdir;    /* custom IFDs are separated from the main ones */
 	union {
@@ -129,7 +129,7 @@ struct tiff {
 	} tif_header;
 	uint16_t               tif_header_size;  /* file's header block and its length */
 	uint32_t               tif_row;          /* current scanline */
-	uint16_t               tif_curdir;       /* current directory (index) */
+    uint32_t               tif_curdir;       /* current directory (index) */
 	uint32_t               tif_curstrip;     /* current strip for read/write */
 	uint64_t               tif_curoff;       /* current offset for read/write */
 	uint64_t               tif_dataoff;      /* current offset for writing dir */
